Deactivate expiry of anti-CSRF tokens
If the tokens expire after a short amount of time, users permanently run into trouble when trying to submit a form. In the worst case, they write a long text and then lose everything due to an expired token. This is obviously a problem.

It also doesn't make a lot of sense from a security perspective. The anti-CSRF tokens are less important than the session ID, so there's no reason for rotating them more often.